requirements.txt
tensorflow
tensorflow-gpu
jupyter
matplotlib==3.3.1
PrettyTable
tf_slim==1.0.0
numpy==1.16.0
pillow==7.0.0
flask==1.1.1
